+++ backend/users/Touir.ExpensesManager.Users/Infrastructure/UsersAppDbContext.cs
@@ -37,6 +37,8 @@ namespace Touir.ExpensesManager.Users.Infrastructure
                 entity.Property(u => u.IsEmailValidated).HasColumnName("USR_IsEmailValidated");
                 entity.Property(u => u.EmailValidationHash).HasColumnName("USR_EmailValidationHash");
                 entity.Property(u => u.IsDisabled).HasColumnName("USR_IsDisabled");
+                entity.Property(u => u.IsDeleted).HasColumnName("USR_IsDeleted").HasDefaultValue(false);
+                entity.Property(u => u.DeletedAt).HasColumnName("USR_DeletedAt");
 
                 entity.HasOne(u => u.CreatedBy)
                       .WithMany()